+++ storybook/storyLoader.js
@@ -4,87 +4,87 @@
 // https://github.com/elderfo/react-native-storybook-loader.git
 
 function loadStories() {
-	require('../source/components/atoms/Box/Box.stories');
-	require('../source/components/atoms/Button/Button.stories');
-	require('../source/components/atoms/Checkbox/Checkbox.stories');
-	require('../source/components/atoms/Fieldset/Fieldset.stories');
-	require('../source/components/atoms/Heading/Heading.stories');
-	require('../source/components/atoms/Icon/Icon.stories');
-	require('../source/components/atoms/Input/Input.stories');
-	require('../source/components/atoms/Label/Label.stories');
-	require('../source/components/atoms/Progressbar/Progressbar.stories');
-	require('../source/components/atoms/RadioButton/RadioButton.stories');
-	require('../source/components/atoms/Select/Select.stories');
-	require('../source/components/atoms/Text/Text.stories');
-	require('../source/components/molecules/BackNavigation/BackNavigation.stories');
-	require('../source/components/molecules/CalendarPicker/CalendarPickerForm.stories');
-	require('../source/components/molecules/Card/Card.stories');
-	require('../source/components/molecules/CheckboxField/CheckboxField.stories');
-	require('../source/components/molecules/DateTimePicker/DateTimePickerForm.stories');
-	require('../source/components/molecules/EditableList/EditableList.stories');
-	require('../source/components/molecules/FormField/FormField.stories');
-	require('../source/components/molecules/GroupedList/GroupedList.stories');
-	require('../source/components/molecules/Header/Header.stories');
-	require('../source/components/molecules/HelpButton/HelpButton.stories');
-	require('../source/components/molecules/ImageUploader/ImageUploader.stories');
-	require('../source/components/molecules/ListItem/Lists.stories');
-	require('../source/components/molecules/Markdown.stories');
-	require('../source/components/molecules/Modal/Modal.stories');
-	require('../source/components/molecules/PersonField/PersonField.stories');
-	require('../source/components/molecules/RadioGroup/RadioGroup.stories');
-	require('../source/components/molecules/RenderCard/RenderCard.stories');
-	require('../source/components/molecules/RepeaterField/RepeaterField.stories');
-	require('../source/components/molecules/ToastNotification/ToastNotifications.stories');
-	require('../source/components/organisms/FormList/FormList.stories');
-	require('../source/components/organisms/Step/Step.stories');
-	require('../source/components/organisms/Step/StepBanner/StepBanner.stories');
-	require('../source/components/organisms/Step/StepDescription/StepDescription.stories');
-	require('../source/components/organisms/Step/StepFooter/StepFooter.stories');
-	require('../source/components/organisms/SummaryList/SummaryList.stories');
-	require('../source/components/organisms/YoutubePopup/YoutubePopup.stories');
-	require('../source/containers/Form/Form.stories');
+  require('../source/components/atoms/Box/Box.stories');
+  require('../source/components/atoms/Button/Button.stories');
+  require('../source/components/atoms/Checkbox/Checkbox.stories');
+  require('../source/components/atoms/Fieldset/Fieldset.stories');
+  require('../source/components/atoms/Heading/Heading.stories');
+  require('../source/components/atoms/Icon/Icon.stories');
+  require('../source/components/atoms/Input/Input.stories');
+  require('../source/components/atoms/Label/Label.stories');
+  require('../source/components/atoms/Progressbar/Progressbar.stories');
+  require('../source/components/atoms/RadioButton/RadioButton.stories');
+  require('../source/components/atoms/Select/Select.stories');
+  require('../source/components/atoms/Text/Text.stories');
+  require('../source/components/molecules/BackNavigation/BackNavigation.stories');
+  require('../source/components/molecules/CalendarPicker/CalendarPickerForm.stories');
+  require('../source/components/molecules/Card/Card.stories');
+  require('../source/components/molecules/CheckboxField/CheckboxField.stories');
+  require('../source/components/molecules/DateTimePicker/DateTimePickerForm.stories');
+  require('../source/components/molecules/EditableList/EditableList.stories');
+  require('../source/components/molecules/FormField/FormField.stories');
+  require('../source/components/molecules/GroupedList/GroupedList.stories');
+  require('../source/components/molecules/Header/Header.stories');
+  require('../source/components/molecules/HelpButton/HelpButton.stories');
+  require('../source/components/molecules/ImageUploader/ImageUploader.stories');
+  require('../source/components/molecules/ListItem/Lists.stories');
+  require('../source/components/molecules/Markdown.stories');
+  require('../source/components/molecules/Modal/Modal.stories');
+  require('../source/components/molecules/PersonField/PersonField.stories');
+  require('../source/components/molecules/RadioGroup/RadioGroup.stories');
+  require('../source/components/molecules/RepeaterField/RepeaterField.stories');
+  require('../source/components/molecules/ToastNotification/ToastNotifications.stories');
+  require('../source/components/organisms/FormList/FormList.stories');
+  require('../source/components/organisms/Step/Step.stories');
+  require('../source/components/organisms/Step/StepBanner/StepBanner.stories');
+  require('../source/components/organisms/Step/StepDescription/StepDescription.stories');
+  require('../source/components/organisms/Step/StepFooter/StepFooter.stories');
+  require('../source/components/organisms/SummaryList/SummaryList.stories');
+  require('../source/components/organisms/YoutubePopup/YoutubePopup.stories');
+  require('../source/containers/Form/Form.stories');
+  require('../source/containers/RenderCard/RenderCard.stories');
 }
 
 const stories = [
-	'../source/components/atoms/Box/Box.stories',
-	'../source/components/atoms/Button/Button.stories',
-	'../source/components/atoms/Checkbox/Checkbox.stories',
-	'../source/components/atoms/Fieldset/Fieldset.stories',
-	'../source/components/atoms/Heading/Heading.stories',
-	'../source/components/atoms/Icon/Icon.stories',
-	'../source/components/atoms/Input/Input.stories',
-	'../source/components/atoms/Label/Label.stories',
-	'../source/components/atoms/Progressbar/Progressbar.stories',
-	'../source/components/atoms/RadioButton/RadioButton.stories',
-	'../source/components/atoms/Select/Select.stories',
-	'../source/components/atoms/Text/Text.stories',
-	'../source/components/molecules/BackNavigation/BackNavigation.stories',
-	'../source/components/molecules/CalendarPicker/CalendarPickerForm.stories',
-	'../source/components/molecules/Card/Card.stories',
-	'../source/components/molecules/CheckboxField/CheckboxField.stories',
-	'../source/components/molecules/DateTimePicker/DateTimePickerForm.stories',
-	'../source/components/molecules/EditableList/EditableList.stories',
-	'../source/components/molecules/FormField/FormField.stories',
-	'../source/components/molecules/GroupedList/GroupedList.stories',
-	'../source/components/molecules/Header/Header.stories',
-	'../source/components/molecules/HelpButton/HelpButton.stories',
-	'../source/components/molecules/ImageUploader/ImageUploader.stories',
-	'../source/components/molecules/ListItem/Lists.stories',
-	'../source/components/molecules/Markdown.stories',
-	'../source/components/molecules/Modal/Modal.stories',
-	'../source/components/molecules/PersonField/PersonField.stories',
-	'../source/components/molecules/RadioGroup/RadioGroup.stories',
-	'../source/components/molecules/RenderCard/RenderCard.stories',
-	'../source/components/molecules/RepeaterField/RepeaterField.stories',
-	'../source/components/molecules/ToastNotification/ToastNotifications.stories',
-	'../source/components/organisms/FormList/FormList.stories',
-	'../source/components/organisms/Step/Step.stories',
-	'../source/components/organisms/Step/StepBanner/StepBanner.stories',
-	'../source/components/organisms/Step/StepDescription/StepDescription.stories',
-	'../source/components/organisms/Step/StepFooter/StepFooter.stories',
-	'../source/components/organisms/SummaryList/SummaryList.stories',
-	'../source/components/organisms/YoutubePopup/YoutubePopup.stories',
-	'../source/containers/Form/Form.stories'
+  '../source/components/atoms/Box/Box.stories',
+  '../source/components/atoms/Button/Button.stories',
+  '../source/components/atoms/Checkbox/Checkbox.stories',
+  '../source/components/atoms/Fieldset/Fieldset.stories',
+  '../source/components/atoms/Heading/Heading.stories',
+  '../source/components/atoms/Icon/Icon.stories',
+  '../source/components/atoms/Input/Input.stories',
+  '../source/components/atoms/Label/Label.stories',
+  '../source/components/atoms/Progressbar/Progressbar.stories',
+  '../source/components/atoms/RadioButton/RadioButton.stories',
+  '../source/components/atoms/Select/Select.stories',
+  '../source/components/atoms/Text/Text.stories',
+  '../source/components/molecules/BackNavigation/BackNavigation.stories',
+  '../source/components/molecules/CalendarPicker/CalendarPickerForm.stories',
+  '../source/components/molecules/Card/Card.stories',
+  '../source/components/molecules/CheckboxField/CheckboxField.stories',
+  '../source/components/molecules/DateTimePicker/DateTimePickerForm.stories',
+  '../source/components/molecules/EditableList/EditableList.stories',
+  '../source/components/molecules/FormField/FormField.stories',
+  '../source/components/molecules/GroupedList/GroupedList.stories',
+  '../source/components/molecules/Header/Header.stories',
+  '../source/components/molecules/HelpButton/HelpButton.stories',
+  '../source/components/molecules/ImageUploader/ImageUploader.stories',
+  '../source/components/molecules/ListItem/Lists.stories',
+  '../source/components/molecules/Markdown.stories',
+  '../source/components/molecules/Modal/Modal.stories',
+  '../source/components/molecules/PersonField/PersonField.stories',
+  '../source/components/molecules/RadioGroup/RadioGroup.stories',
+  '../source/components/molecules/RepeaterField/RepeaterField.stories',
+  '../source/components/molecules/ToastNotification/ToastNotifications.stories',
+  '../source/components/organisms/FormList/FormList.stories',
+  '../source/components/organisms/Step/Step.stories',
+  '../source/components/organisms/Step/StepBanner/StepBanner.stories',
+  '../source/components/organisms/Step/StepDescription/StepDescription.stories',
+  '../source/components/organisms/Step/StepFooter/StepFooter.stories',
+  '../source/components/organisms/SummaryList/SummaryList.stories',
+  '../source/components/organisms/YoutubePopup/YoutubePopup.stories',
+  '../source/containers/Form/Form.stories',
+  '../source/containers/RenderCard/RenderCard.stories',
 ];
 
 module.exports = {
